Bugfix in the typechecker.
Memoization was incorrect. We need to (re-)compute the replacement for
polymorphic operators even if we already know the type of the
expression. So the memo must be indexed by the AST Node ID, and not just
the expression and its context. (Because if we do the latter, we may
just look at an expression and think we know the type of it, but later
when we look up the reified op in the PolymorphicTypeRewriter pass, we
won't have one.

git-svn-id: https://repo.eecs.berkeley.edu/svn/users/spramod/uclid5/trunk@87 79859994-26b0-47d7-b2a6-c933f2ed28e0